requirements.txt
# Core Libraries
gspread
pandas
openpyxl
google-auth-oauthlib
google-auth-httplib2

# Web Automation
playwright

# Scheduling
schedule

# Telegram Bot
python-telegram-bot

# Linter
flake8
